Upgrading to svn revision 6511

+++ plugins/PrivacyManager/LogDataPurger.php
@@ -18,6 +18,11 @@ class Piwik_PrivacyManager_LogDataPurger
 	const TEMP_TABLE_NAME = 'tmp_log_actions_to_keep';
 	
 	/**
+	 * The max set of rows each table scan select should query at one time.
+	 */
+	public static $selectSegmentSize = 100000;
+	
+	/**
 	 * The number of days after which log entries are considered old.
 	 */
 	private $deleteLogsOlderThan;
@@ -73,8 +78,16 @@ class Piwik_PrivacyManager_LogDataPurger
 			}
 		}
 		
-		// delete unused actions from the log_action table
-		$this->purgeUnusedLogActions();
+		// delete unused actions from the log_action table (but only if we can lock tables)
+		if (Piwik::isLockPrivilegeGranted())
+		{
+			$this->purgeUnusedLogActions();
+		}
+		else
+		{
+			$logMessage = get_class($this).": LOCK TABLES privilege not granted; skipping unused actions purge";
+			Piwik::log($logMessage);
+		}
 		
 		// optimize table overhead after deletion
 		Piwik_OptimizeTables($logTables);
@@ -142,14 +155,26 @@ class Piwik_PrivacyManager_LogDataPurger
 	 */
 	private function getDeleteIdVisitOffset()
 	{
+		$logVisit = Piwik_Common::prefixTable("log_visit");
+		
+		// get max idvisit
+		$maxIdVisit = Piwik_FetchOne("SELECT MAX(idvisit) FROM $logVisit");
+		if (empty($maxIdVisit))
+		{
+			return false;
+		}
+		
+		// select highest idvisit to delete from
 		$dateStart = Piwik_Date::factory("today")->subDay($this->deleteLogsOlderThan);
 		$sql = "SELECT idvisit
-		          FROM ".Piwik_Common::prefixTable("log_visit")."
-		         WHERE '".$dateStart->toString('Y-m-d H:i:s')."' > visit_last_action_time AND idvisit > 0
+		          FROM $logVisit
+		         WHERE '".$dateStart->toString('Y-m-d H:i:s')."' > visit_last_action_time
+		           AND idvisit <= ?
+		           AND idvisit > ?
 		      ORDER BY idvisit DESC
 		         LIMIT 1";
-
-		return Piwik_FetchOne($sql);
+		
+		return Piwik_SegmentedFetchFirst($sql, $maxIdVisit, 0, -self::$selectSegmentSize);
 	}
 
 	private function getLogTableDeleteCount( $table, $maxIdVisit )
@@ -261,14 +286,17 @@ class Piwik_PrivacyManager_LogDataPurger
 	}
 	
 	// let's hardcode, since these are not dynamically created tables
-	// exclude piwik_log_action since it is a lookup table
 	public static function getDeleteTableLogTables()
 	{
-		return Piwik_Common::prefixTables('log_conversion',
-										  'log_link_visit_action',
-										  'log_visit',
-										  'log_conversion_item',
-										  'log_action');
+		$result = Piwik_Common::prefixTables('log_conversion',
+											 'log_link_visit_action',
+											 'log_visit',
+											 'log_conversion_item');
+		if (Piwik::isLockPrivilegeGranted())
+		{
+			$result[] = Piwik_Common::prefixTable('log_action');
+		}
+		return $result;
 	}
 
 	/**
